archive: refactor to put code into 'commands/archive_cmd.py' module.
